requirements.txt
discord-py-interactions==4.3.1
interactions-files==1.1.3
requests==2.24.0
bs4
psycopg2
disputils==0.2.0
dblpy